+++ example/lib/src/self_registered_service.dart
@@ -1,0 +1,16 @@
+import './transport.dart';
+
+abstract class SelfRegisteredService {
+  void sayHello();
+}
+
+class MySelfRegisteredService implements SelfRegisteredService {
+  final Transport _transport;
+
+  MySelfRegisteredService(this._transport);
+
+  @override
+  void sayHello() {
+    _transport.transferData('Hello from MySelfRegisteredService');
+  }
+}