+++ lib/src/builder/generator/service_provider/methods/methods.dart
@@ -2,6 +2,7 @@ export 'boot.dart';
 export 'constructor.dart';
 export 'ensure_booted.dart';
 export 'has.dart';
+export 'register.dart';
 export 'resolve.dart';
 export 'resolve_or_get_parameter.dart';
 export 'try_resolve.dart';