+++ test/unit/service_descriptor_test.dart
@@ -1,9 +1,12 @@
 import 'package:catalyst_builder/catalyst_builder.dart';
 import 'package:test/test.dart';
 
-void main(){
+void main() {
   test('ServiceDescriptor constructor', () {
-    var descriptor = ServiceDescriptor<int>(const Service(exposeAs: int), () => 1337);
+    var descriptor = ServiceDescriptor<int>(
+      const Service(exposeAs: int),
+      () => 1337,
+    );
     expect(descriptor.service.exposeAs, equals(int));
     expect(descriptor.produce(), equals(1337));
   });